Automatic git commit store in trial dir
This feature is very useful if you need to reconstruct the exact state
of a iteration.

If it is detected that the a git repo is present, it will store the
commit hash of the HEAD in the git.txt in the iteration directory. So if in
the future you need to restore the exact state of the benchmark, just
do a git checkout of that hash.

It is needed that there are no untracked files or dirty changes in that
case; if there is.. an error will be shown and SSO will exit.

If you don't have git installed or are not working in a git repo, then
nothing will happen.