Minor improvements
The file containing the HEAD is now called HEAD.

If you very quickly create Iterations, it could be you run into a
directory already exists error. So a check is done and there is a
short sleep and a new attempt is made. This is because the directory
name contains the time.