Add RAID class and Cassandra support for RAID
Add a RAID class which sets up a RAID of chosen devices (for example
multiple NVMe disks on i3 instances), performs some tuning of them
and sets up a filesystem (by default XFS). A RAID level is configurable
(by default RAID-0).

A support of RAID is added to Cassandra class, to set up a RAID array
with NVMe devices (tested on i3 instances with multiple disks).
Cassandra is installed on this array, so that the data files are stored
there.

Note that some NVMe-specific and i3-specific constants are hardcoded
in Cassandra class. In the future this code could be made more generic
to accommodate different setups (not only i3 instances).