Improved ssh output.
Instead of getting a direct copy of the output of the remote
machine, each line gets spaces followed by the ip and then
the content.

This makes it a lot easier to see what is going on when multiple
ssh sessions are running in parallel (e.g. when running benchmarks).

Because of the identation, it is also easier to see the structure
in the output.