generate_benchmark resourceid option.
Before this change, when you generate 2 benchmarks,
they will share the same resource names which can lead
to problems when running on EC2 if you want to run
them in parallel.

So a new option has been added to provide a custom
resource-id when generating a benchmark.

By default the username is used for the resource id.