Iteration ignore git by default
The iteration object will check if a git repo exist and if so,
it expects this repo to be clean.

This can lead to a problem when the benchmark is part of some
git repo at a higher level.

So by default, git is ignored and no reference to the commit
is stored in the iteration directory.