Various error handling fixes.
This fixes a problem with e.g. insertion. If insertion fails, you will
see error logging in the console. But the error handling was broken and
it makes you believe there was no problem because it says data the
insertion of the data completed.

There are 2 causes:
- The cassandra-stress call lost the error code due to the 'tee'
  command.
- The future didn't handle the exception correctly. If an exception
 is set, instead of propagating the exception to the caller, it
 is interpreted as regular value which in most cases is ignored.

Both of these issues have been resolved in this PR.